% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/austin_weather.R
\docType{data}
\name{austin_weather}
\alias{austin_weather}
\title{Austin Weather Data}
\format{
A data frame with 365 rows and 4 variables:
\describe{
  \item{month}{Month of the Year}
  \item{day}{Day of the Year}
  \item{yearhome}{Year}
  \item{temp}{Temp}
}
}
\usage{
austin_weather
}
\description{
Austin Weather Data for 2019
}
\keyword{datasets}
