% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem.R
\name{sem}
\alias{sem}
\title{Get Standard Error of the Mean}
\usage{
sem(x)
}
\arguments{
\item{x}{a vector of numbers}
}
\value{
the standard error of the mean
}
\description{
Get Standard Error of the Mean
}
\examples{
sem(c(1,3,5,8,11))
}
